Moved each view to it's own file. Added sun_ prefixes.
