+++ CppSundial/src/sun/sun.cpp
@@ -1,0 +1,14 @@
+#define SUN_IMPL
+#include "sun/views.h"
+#include "sun.h"
+#include "sun/sunUtils.h"
+
+#include "sun/graph.h"
+#include "sun/controls.h"
+#include "sun/paths.h"
+#include "sun/field.h"
+#include "sun/network.h"
+#include "sun/swerve.h"
+#include "sun/powerIndicators.h"
+
+SunGlobs globs;
